Fix questions model
